Fix: category 폴더 분류
2024.10.24
