+++ controller/space.js
@@ -1,7 +1,7 @@
 import db from '../models/index.js';
 import multer from 'multer';
 import jwt from 'jsonwebtoken';
-const { Category } = db;
+
 
 const storage = multer.diskStorage({
     destination: (req, res, cb) => {
@@ -175,35 +175,3 @@ export const addNewSpace = async (req, res) => {
 //ANCHOR - 이미지업로드
 export const uploadSpaceImage = upload.array('image', 10);
 
-//카테고리 추가
-export const addCategory = async (req, res) => {
-    const { categoryName, pId } = req.body;
-    if (!categoryName) {
-        return res
-            .status(400)
-            .json({ result: false, message: '카테고리 이름을 입력해주세요.' });
-    }
-    try {
-        const duplicationCategory = await Category.findOne({
-            where: { categoryName },
-        });
-
-        if (duplicationCategory) {
-            return res.status(401).json({
-                result: false,
-                message: ' 이미 존재하는 카테고리 이름입니다',
-            });
-        }
-        const newCategory = await Category.create({
-            categoryName,
-            pId: pId || null,
-        });
-        res.status(201).json({ result: true, data: newCategory });
-    } catch (error) {
-        res.status(500).json({
-            result: false,
-            message: '카테고리 저장 중 오류가 발생했습니다.',
-            error,
-        });
-    }
-};
