+++ router/space.js
@@ -1,8 +1,7 @@
 import express from 'express';
-import { addCategory, addNewSpace, uploadSpaceImage } from '../controller/space.js';
+import { addNewSpace, uploadSpaceImage } from '../controller/space.js';
 
 const router = express.Router();
 
 router.post('/add-new-space', uploadSpaceImage, addNewSpace);
-router.post('/add-category', addCategory)
 export default router;
